+++ src/ArduinoHA.h
@@ -10,6 +10,7 @@
 #include "device-types/HACover.h"
 #include "device-types/HADeviceTracker.h"
 #include "device-types/HALock.h"
+#include "device-types/HASwitch.h"
 #include "device-types/HATagScanner.h"
 
 #ifdef ARDUINOHA_TEST